Fully use patterns for flag swapping
Previously swapflags() only partially used patterns for flag swapping,
and it was kind of like a weird amalgamation between patterns and other
string processing mechanisms.

It's shorter and a bit more readable than a bunch of string
concatenation (where you'd have to decipher all the indices being passed
to :sub()).

Here's how it works: Assume `command` is "ifflag", `flag1` is 76, and
`flag2` is 34.

Then the gsub will be to match "^(ifflag[%(,%)])76" and replace it with
"%134", which you should read as "%1" concatenated to "34".

The caret means to match only at the start of the line. The parentheses
around "ifflag[%(,%)]" are to match it as a group (which will be group
1) so I can repeat it in the replacement string with "%1". "%(,%)" is
surrounded in square brackets so that we'll match either "%(", ",", or
"%)", and the percent signs are to escape the parentheses.