Ignore leading zeroes in swapflags()
Although, like spaces, they will be ignored in detection but not
preserved in the swapping.

The dash means to match as few zeroes as necessary. I could have used
the asterisk, which is the same thing but means to match as many zeroes
as possible ("greedy"), but I feel like if there's any difference
between the two, the non-greedy one is safer.