Improved filefunc API and variables
The full path to the levels folder used to be the only path that was
kept track of. Since the graphics folder was a little bit of interest
as well, it was kind of an afterthought, and it was accessed based on
the levels folder path.

However, with the assets viewers, we'll need to be able to do more with
the other folders, including listing files in them. Ved will now keep
more path variables, making stuff more elegant. `getlevelsfolder` no
longer returns the levels folder as the second argument, instead it
sets global variables directly (which was happening anyway).
`listfiles` has been renamed `listlevelfiles`.

The filefunc `readimage` has been generalized to be `readfile`, and no
longer takes the path to the levels folder as an argument. There's now
also a set of filefuncs that can be used to write to a file in separate
operations, which is used for writing music files.

I also added more functionality to the C filesystem library for Linux
and macOS, mainly in preparation for listing non-level files in other
folders than the levels folder.

[pre13]