Recenter window if non-1x or smaller screen
At the start, LO"VE is friendly and centers our 896x480 window for us.
However after that it doesn't do anything else. So if the scale isn't 1x
or we are using the smaller window size, then the window will be
off-center.

To fix this, do it when we resize the window for scale settings. We
unfortunately have to infer the size of the display we're on from the
position and width alone, which can be done, but for all intents and
purposes is an ugly hack until LO"VE implements functions to get screen
width and height (which I can't seem to find).

Also there's now a new line of console output printing the size of the
display detected in case any issues arise from this inference.