Add support for tileset 0 (Space Station) tilecol -1
This adds support for the blue unpatterned Space Station tiles
immediately to the left of tilecol 0 in tiles.png.

This tilecol is mostly well-behaved, except moving-type and disappearing
platforms are invisible, conveyors facing left animate using tiles 2-5
in tiles.png, conveyors facing right use the animation of tileset 0
tilecol 0 facing left (so they're backwards), and in the in-game editor
enemies are red but if you actually play it in-game they're green.

Other than that, though, they use the blue backing, and have a gray warp
background.

Since Ved's entcolour support is limited to specifying only a single RGB
color for EVERYTHING, I decided to just make it the same gray as the
Warp Zone gray for consistency. Also, I'm not sure what colorid means,
but I set it to 10199 since I couldn't really put a negative in there.
And I'm also wondering what tilesetid means.

This commit adds the following:
 - The tiles of the tilecol themselves
 - Support for cycling and revcycling to tileset 0 tilecol -1
 - An updated level check so Ved won't complain if there's a tileset 0
   tilecol -1 in the level upon level load

Closes http://ved.idea.informer.com/proj/?ia=110896
Closes #15