Released Ved 1.5.1b (Windows+Source)
This is exactly the same as the latest officially released
version 1.5.1, except that it fixes the crash that some people get at
startup.
(so it's 1.5.1 plus commit f24436b3577fcdd91b6ce3b6523a682532e3d2d7)

This problem was already solved for the time being by pointing people
to a replacement .exe and .love for 1.5.2-pre03, a few commits ahead,
but some people missed that, and at the very least it's a barrier to
getting Ved to work if you encounter this crash.

So, I'm replacing download links for 1.5.1 with 1.5.1b, which for
Windows is now a full zip with all the DLLs instead of just the .exe.

Releasing 1.6.0 (which, for the record, is the next version after
1.5.1, since 1.5.2 got a version number bump before release) would fix
this as well, of course, but that's a pretty big version and will take
more time to finish than I'm comfortable with to leave this barrier in
place.