Fix crash from exiting int.sc cleared with Ctrl+D
What was happening was Ctrl+D actually removed every line from the
script, even though it looks like there's a blank line if you clear it.
Typing something would add a line back. Pressing ENTER would add it
back, too, along with creating a new line.

But if there were literally no lines in the script, then upon exiting,
if the script was a loadscript or say(-1) internal script, Ved would
attempt to fetch the last line of the script to see if it was
`iftrinkets,0,`, `customiftrinkets,0,` or `loadscript`. However, since
the last line was nonexistent, Ved crashed.

This commit makes it so that if you press Ctrl+D or Ctrl+Shift+D in a
script with only one line, Ved won't literally remove the line, and
instead will simply blank it, fixing the bug.