Fix ENTER on fopen error dialog w/o animation opening it again
If you pressed ENTER on an fopen error dialog (if you tabbed over to a
level), the dialog would close but then come right back up.

This commit fixes it so that it will now close properly instead of
opening again.

To do this, a new global variable called `returnpressed` was introduced.
It might look confusing if you're still thinking of "return" as in "go
back" but we call the ENTER key the Return key.