Restore Delete key remove-from-recently-opened functionality
In a BRILLIANT display of violating Chesterton's fence (to elaborate,
the general principle to prevent shooting yourself in the foot invented
by G.K. Chesterton, where if you see a mysterious fence and do not
understand its purpose, you should not remove it until you do understand
its purpose, the fence standing in as a tangible object for customs,
traditions, and things that are already there), earlier in
75a1a450613da203850d4bd6b0ad2b9f00430b9e, I wondered why the Delete key
did not reset the tabbing in level select to focus on level input like
Backspace and typing anything did, and concluded it was part of the
general asymmetry in every input field where Delete didn't have the same
powers as a reverse Backspace.

And then I saw that Delete lets you remove levels from the list of
recently opened levels list.

So now I have put a bigger fence in the form of a comment telling the
reader why resetting tab selection with Delete is not such a good idea,
along with commenting out the code (so the reader knows exactly which
code is not a good idea).