Fix crash from dragging 1_10 onto the main canvas without touching it
1_10 is the Potato for Doing Things that are Magical.

If you hadn't put your mouse on the main canvas yet, then clicked on the
potato and then dragged it onto the main canvas, Ved would crash.

No, this diff only points to the part of the code where the potato
subtool is actually handled. It does not point to how to unlock it. So
don't look here, it's a waste of your time.