Fix performance bottleneck in level diffs
Now that roomdata and levelmetadata accesses are proxied through a
function, they're a LOT more expensive now, as seen in this for-loop
where roomdata_get() was being called on the same exact room every time.
The fix was to move the roomdata_get() to the top of the for-loop,
before it.