Added graphics viewer
The graphics viewer allows you to open an image file (normally ones in
your graphics folder, but you can open anything you like that LÖVE
supports from anywhere in your filesystem you have access to).

https://cdn.discordapp.com/attachments/403718283855134720/597035501257752576/unknown.png

Viewing the image, you can view tile or sprite numbers based on the
grid setting, which can be 1, 8 or 32 pixels (or disabled). This will
be set automatically based on what the filename starts with, so opening
"tiles.png" or even something like "tiles2-old.png" will set the grid
to 8 pixels, and a file like sprites.png will be opened with a grid for
tiles of 32x32 pixels. When I say "grid", it's not a visual grid of
lines, it's mainly so you can hover over tiles. And you can always
change this setting.

1.6.0 is getting pretty close to release!

[pre18]