Clarified interface for int.sc mode methods
TL;DR: Earlier say(-1) was added as a second method for int.sc mode,
and the int.sc mode button had three states instead of being on/off.
I think the button labels are too confusing though, so I split the
button into two separate buttons that hopefully makes it clearer,
while this new method hasn't yet been officially released (in 1.6.0).

!14 adds a second internal scripting mode method, referred to by my
request as "say(-1) int.sc" instead of "C.bars int.sc". So as of 1.6.0,
the int.sc mode button would be a cycle button instead of a toggle,
which cycles as follows:

Int.sc is off > Loadscript int.sc > say(-1) int.sc

Despite renaming the methods, I think that control is kind of unclear -
everyone is used to just int.sc being either off or on, and requiring a
loadscript. It's not clear to everyone which of the two modes is the
one that existed before, so I feel there needs to be some kind of hint
that "Loadscript int.sc" is the good-old "Int.sc is on" and
"say(-1) int.sc" is the alternative option that sometimes gets
deprecated ("the old method of internal scripting", "say(-1) is
outdated").

(This deprecation is less justified today because nobody specifically
aims to support VVVVVV 2.0 anymore, so we can now use squeak(off) to
get rid of the annoying forced terminal sound or squeak that the
loadscript method didn't have in the first place. Which Ved's new
say(-1) method support does.)

I changed the 3-tier cycle button to two buttons: one on/off button,
and a button to change the method. When int.sc mode is off, the second
button is not shown. Turning int.sc mode on will always default to
"Loadscript int.sc", which is what "Int.sc is on" was before. When the
method is set to "say(-1) int.sc", the second button is yellow, as kind
of a visual hint that it's set to the "alternative" mode. The ctrl+I
and ctrl+shift+I shortcuts remain as they were, and cycle as before.
I removed the button to create a new script from the script editor, to
make room for the extra int.sc method button. (To be clear, the new
script button in the scripts list remains, of course.) I've said before
that the script editor kinda needs an interface overhaul, since space
is tight there. Therefore, this two-button interface may change later.

[pre21]