Add setting s.allowbiggerthan20x20
Ved seems to only write the new config option when it writes its config
(i.e. when you go into Ved options and click OK), while I was expecting
it to simply pop up when you ran the new version. I don't know what to
do about that, or if there's something I missed where I have to add the
new config option to a list of things to check and write on startup.
Also, especially since newer config options can be removed if you write
the settings.lua file with an older version of Ved, which is a bit
annoying.