Check s.allowbiggerthan20x20 in loading and resizing levels
A new devstring has been added for the failed check if a map is bigger
than 20x20. And the "maximum size of a level is 20 by 20" dialog is back
now.

Also, if the map width or height is bigger than 20x20 now (and
s.allowbiggerthan20x20 is false), the offending dimension in question is
set to 20 instead of 1, finally.