Make dialogs go from maxlevelsize -> biggersize
Now, if you have s.allowbiggerthan20x20 set to true and you make a map
smaller than 20x20 to be bigger than 20x20, the max level size will pop
up, but it will have Override and Don't Override buttons. Clicking
Override will lead to the usual confirmation dialog.

I made the non-Override button Don't Override instead of OK or Cancel or
or anything else because I didn't want it to be ambiguous about which
keyboard shortcut to press (in the case of OK, since ENTER is used to
choose OK but here ENTER is ambiguous as to if it's OK to the override
or OK to not overriding) or if choosing Cancel would cancel the map
resize entirely, and all other buttons don't fit the wording.

Ved will also defer setting the map size until the end of the sequence
of dialogs, instead of pouncing on it for each dialog. This should
prevent border lines from changing if you're in a smaller than 20x20
(but not 20x20) map and you're on an edge room.