Added a few more shortcuts, added Info Teddy to the credits
In the levels list, ctrl+A now goes to the Assets screen.
In the script editor, ctrl+A copies the entire script (like the "Copy
script" button), and ctrl+C copies the current line. Also, I added a
missing ctrl+right F9 shortcut hint to the "Go to"/"Create" script
button in the script editor (when you're on a line with
ifflag/iftrinkets/etc)

I have yet to add all of this to the shortcuts help page, and speaking
of adding shortcuts to the help page, see #55; but first I want to
change the way the help articles can be translated (it's a nightmare to
change stuff in the help pages at the moment, when every article is ONE
string). But before I do that, I want to get 1.6.0 out.

What I have done is add Info Teddy to the credits as a code
contributor. He has made lots of contributions in the past months:
https://gitgud.io/Dav999/ved/merge_requests?scope=all&state=merged&author_username=infoteddy

[pre25]