Reduced chance of incorrect split in int.sc mode
In short, lines above lines within textboxes starting with `text` would
be marked safe to split, thus possibly breaking up a text box if it
happened to be at a splitting point. Now it's still not that
waterproof, but it does now require at least 4 argument separators to
be on the line, as well as an argument separator following the `text`.

Also, spaces are now accounted for when determining safe splitting
points.

[pre02]