Move roomname up 1px and make background 10px high
In Ved, since the main canvas is double-scaled, this actually translates
to moving the roomname up 2 pixels and making the roomname background 20
pixels high.

This change is because in VVVVVV, roomnames are 1 pixel away from the
bottom of the screen instead of touching it directly like Ved, and
roomname backgrounds are 10 pixels high instead of 8 pixels.

Note that I didn't want to mess with the line height, so for the extra
lines while editing a roomname that are against the red background, I
could've made it favor the bottom line (so the last red line is touching
the gray), or the other lines (so the bottom line is touching the red).
Both look pretty ugly to one of the gray/red background and the lines on
each. I decided instead to make it so there were 2 pixels equally
distributed to both backgrounds, meaning there's a 1 pixel gap until the
gray-background line touches the red and a 1 pixel gap until the
red-background line touches the gray. This results in an uneven height
for both backgrounds, but whatever, Ved already deals with 1 pixel
lengths on the main canvas anyway (e.g. the 1-pixel border lines).