Add s.opaqueroomnamebackground settings.lua option
So you can make the roomname background opaque like it is in VVVVVV.
This will only apply to the black roomname background, e.g. when you are
not editing the roomname (gray background, red background added if you
have an overflowing roomname) and when there's no temporary roomname to
signal an event (yellow background).

Note that border lines behind the roomname background will not show up
if it is opaque. This is also how it is in the in-game editor. Not that
we should follow the mistakes of the in-game editor, but I think if I
made it so that the adjacent room lines drew on top of the roomname, it
could lead to confusion like people assuming that the roomnames are
somehow "underlined", for example. And besides, the point of an opaque
roomname background is so you aren't able to see what's behind it, which
is how it is in-game, so we shouldn't put things behind the roomname in
front of the roomname.

The entity debug tooltip will be covered by the opaque background, but
then again the entity debug tooltip will go offscreen if the entity is
close to the bottom or right edge, instead of snapping to the edge or
moving to the other side, and the tooltip doesn't have a background of
its own so the text can be indistinguishable from white if it goes over
white pixels, so the issue that it goes behind the opaque roomname
background can be dealt with at the same time all those other issues are
dealt with.