Display roomtext on the last 2 rows above roomname bg
This is how it is in-game, but it doesn't show it that way in the
in-game editor, but it does show it that way if you playtest with the
in-game editor.

This is done in a careful way to avoid double-rendering, because later
when translucency in font16 is fixed I won't have to worry about the
translucency stacking due to double-rendering. Also, interaction code is
handled along with rendering an entity and I don't think anything bad
will happen if I run the interaction code twice, but just to be safe I
won't.