Stop input when going to state 15 from 3 or 6 via F1
I was in debug mode, and I kept pressing F1 to go to the help pages to
test the F9 hotkeys, and I noticed that it still said "TAKING INPUT".
Then I typed some stuff in, and lo and behold they showed up when I
returned to level select. So now, they won't.