Add F9 hotkeys for dialog buttons
Note that the window border is drawn after the buttons, thus the hotkey
of the right button gets cut off a little bit. But this will be fixed in
the next commit.

These shortcuts are now shown when you press F9:
 - Enter for OK
 - Esc for Cancel
 - Y for Yes
 - N for No
 - D for Discard
 - S for Save

Additionally, these will only show up for the topmost dialog, to prevent
confusion that pressing any of these would close ALL dialogs at once or
something.