Make gotoroom script context wrap around
In game, if you were to execute `gotoroom,10,10` in a 5x5 map, you would
end up in (1,1) (1-indexed) because the game would wrap around your
coordinates. Therefore, it's much better to display the wrapped-around
room than it is to display the coordinates unfiltered from the command.

The jury's still out on whether showcoordinates/hidecoordinates wraps
coordinates around. From my limited testing, it's a very weird pattern
and I don't yet understand it. So, only gotoroom commands are subjected
to this for now. ifexplored seems to operate the same way as
showcoordinates/hidecoordinates, too, even though ifexplored is not yet
added to any script contexts because of limited UI space/poor UI.