Fix square brackets/parentheses in URL in translation tools readme
The good ol' classic Markdown URL syntax is prone to people mixing up
the square brackets and the parentheses. Which goes where?! Do the
square brackets go with the label or do they go with the URL?!?!?!?!

The square brackets go with the label and the parentheses go with the
URL. My on-the-spot mnemonic is to think of how you'd usually embed a
URL in plaintext - with parentheses after something, like this.

    They said if I go on this site again (reddit.com) they would...

As you can see, no one embeds URLs with square brackets. And whoever
DOES do that, let me formally declare them as a weird person who
probably lives off in a hermit hut in the middle of the mountains or
something.

Anyway yeah the URL is fixed.