Allow platv outside 0-8
I didn't have to make the platv button an input field. I just had to
make it spawn a DIALOG with an input field.

Anyway, this adds devstrings PLATVCHANGE_TITLE, PLATVCHANGE_MSG, and
PLATVCHANGE_INVALID, for, you know, the dialog.

Also, since there's two UNRE events relating to platv now (one with the
dialog, and one where you reset it to 4 by right-clicking), the UNRE
event with the reset to 4 is now labelled "PLATV (reset to 4)".

Closes #37.