Make platv use int_control()
int_control() is the function that makes an increment/decrement
interface, as seen in the options menu for the FPS limiter and number of
backups. Here, it is restricted to platv 0 through 8.

The button to open the dialog and input unrestricted platv is still
there, it's just on the number now. As well, right-clicking in the WHOLE
space (including the label and plus and minus signs) will reset platv to
4.