Account for German and macOS in tinynumbers
This means that if you use the German language, CTRL in hotkeys will now
be STRG. If you are on macOS, CTRL will then be the symbol for CMD. I
haven't actually tested if it works on macOS, but I assume
`love.system.getOS() == "OS X"` is good enough proof that it will. And
if you are a German macOS user, you should know that CMD takes
precedence over STRG, just like it does elsewhere.