Fixed "Copy" buttons in scripts and level notes if cursor is in line
If there was text to the right of the cursor in the script editor,
copying the entire script would not include that part to the right of
the cursor. Same for the button in the level notes editor, but clicking
that button would also move the cursor... I'll mention #86 here. Also,
if we refactor this system, it'd be nice to also stop using generic
input variables when editing multiline things like scripts or level
notes, and edit those more directly.

[pre09]