Opening vvvvvvmusic.vvv or mmmmmm.vvv from outside VVVVVV folder fixed
Opening those files from the VVVVVV folder is special, since it'll just
use the corresponding already-opened vvvvvvmusic.vvv or mmmmmm.vvv from
the two music player screens. However, in two places, I forgot to
account for those filenames outside the VVVVVV folder not being files
inside the VVVVVV folder, because I was only checking the filename
there. Oops. (It did actually load the file in correctly but that was
inaccessible)

[pre11]