Fix main editor L hotkey not showing editorscreenshot
If you pressed L instead of clicking on the folder icon, the
editorscreenshot wouldn't show up in the levels list.

Or rather, it DID show up, but it would be a screenshot of the levels
list, instead of the main editor, which isn't what we want.

Apparently, it's a problem that we switch to state 6 in
love.keypressed() instead of love.draw() like clicking on the folder
icon does. Therefore, this commit will make it so that pressing L
instead just defers the act of switching to state 6 until the next
love.draw(), instead of doing it in love.keypressed().