Make too long filenames scroll in level select
Just like how if the title of the level in the levels list is too long,
it gets cut off with an arrow and will scroll if you tab over to it, or
hover your mouse on it.

This means you can no longer have an ugly visual where the filename
overlaps the title.

The cutoff width (50 characters) also applies to folders, especially in
the Backups screen, even though the title isn't displayed for folders
and there's plenty of space for the filename to take up. It would be a
lot more messy to change the width for folders, and anyway I feel like
it's a good thing that long folder names are cut off, too, so you don't
have to move your eyes across the entirety of the window to read it, and
if you weren't reading it, getting the gist that it's very long is good
enough.