Added temporary workaround for LÖVE 0.9.x lack of Font:setFallbacks
So, we opted to use that for hotkey glyph localization and for showing
the cmd symbol on Mac, but last-minute I realized Font:setFallbacks is
not yet supported on LÖVE 0.9.x, so that would cause a crash if you're
not on at least LÖVE 0.10. I therefore made a temporary version of
tinynumbersfont.png that includes the CTRL glyph, until I can think of
a way to handle this better.

[pre17]