Fix '/' to first script dirtying
If you pressed the '/' (slash) hotkey to go to the first script in the
script list (well, technically the last script as Ved shows the scripts
in descending order), Ved would dirty the level, even though it doesn't
dirty it if you open the script with the mouse, or even dirty it when
you open any other script, even though opening the first script wouldn't
change the script order but opening any other script would!

Basically, Ved temporarily set nodialog to be false to mark when it
should not count characters as typing, which is basically when context
changes. It already does this for the 'e' hotkey to not type an E when
using that hotkey to edit the roomname, and it does this for the '/'
hotkey in order to not type a '/' or a '?' in the script when those
hotkeys are used.

However, while it won't type the '/' or '?', it'll act as if you did
anyway, and then dirty the level. So now I've fixed it so it doesn't do
that.