Show proper spike hitboxes when holding down J
It's very simple - vertical spikes have a hitbox of 8 by 4 pixels.
Spikes facing upwards have this hitbox on their lower half, while spikes
facing downwards have this hitbox on their upper half.

What is less simple is how to display this. Without deviating from the
"X" style, I could've gone with simply squashing the X, making two
smaller Xs, or rendering only half an X. Squashing the X would clash
with the other Xs due to having a different aspect ratio than the rest
of them, two smaller Xs would make it look like there's somehow two
smaller hitboxes instead of just one half-hitbox, and half an X would
simply look like the outline of a spike, but it wouldn't be really
consistent with the rest of the "X" style and it doesn't really fill the
hitbox.

In the end I went with squashing the X down to half-size, as that
conveys the fact that it's a half-hitbox the best, even if its aspect
ratio will clash with the other Xs (also it looks a bit thicker and more
stretched).