Make subtool scrolling work with Shift+</>
Using the scrollwheel by default scrolls through the subtools. Using <
and > by default scrolls through the main tools. You could use Ctrl or
Shift to make the scrollwheel scroll through the main tools, but
previous to this commit, you could use ONLY Ctrl to make < and > scroll
through the subtools, which was a bit inconsistent.

So I've made it so Shift+</> will work to scroll through the subtools as
well.

(Note that </> actually means "',' (comma) or '.' (period)", whose shift
modifiers just happen to be '<' and '>'.)