Reverse font stack order
Previously, it went like "All the chars except the interchangeable ones
have been removed -> The different versions of chars -> The default
versions of said chars", which caused a few issues as LÖVE 0.9.x didn't
support fallbacks, and so had to use a version where all the chars
including the default chars was in place, which was basically
duplicating all the chars.

Now, the stack is like this: "Nothing -> The different versions of chars
-> All the chars AND the default versions of chars". It makes it easier
because for LÖVE 0.9.x we can just make it point to the last font in the
stack instead of making a new font image entirely.

Closes #88.