Add script reorderability
There are now up- and down- arrows on the sides of the script name
buttons, which have been shortened accordingly. Of course, if you're
using a custom font.png then those arrows won't be fancy Unicode and
will instead be '^' and 'V'.

When you click on them, your mouse moves along with the button when the
script moves to its new place.

Closes #54 and ?ia=115896