Add button to list all crewmates
The custom level type, the edentity type, the rescuable one. Not the
createcrewman or createentity one.

If you hold the edentity crewmate tool, there will be a button on the
right panel that you can press, to create a dialog that will list the
coordinates of the rooms containing every crewmate in the level.

This is just to be basically symmetrical with the find all trinkets
button.