Add room shifting on the map screen
From http://ved.idea.informer.com/proj/?ia=103333

Instead of going with the clunky keyboard-based interface suggested by
StillSpelledViridian, I went with clicky arrow thingies. I also added a
checkbox to update scripts as well. While it may look like it applies as
well to "Swap rooms", unfortunately, it doesn't, because that feature
was never programmed with scripts in mind.

Shifting rooms isn't undoable because it modifies scripts and editing
scripts isn't undoable. I could make it so that it's undoable if you
uncheck "Update spts." but that'd just be plain inconsistent, which I
don't want.

Closes http://ved.idea.informer.com/proj/?ia=103333