Make flag swapping work with no second arg separator
Previously, I made it work only with '<command>,<flag>,' because if I
didn't, it would (if you swapped 1 and 2) replace '<command>,10' with
'<command>,20', which would be bad. But now we can use '$' (the dollar
sign) to make sure that if the flag is at the end of the line, we can
replace it without worry that we're actually replacing 10 instead of 1.