Prevent more simultaneous roomname+something else events
Left-clicking, right-clicking, or middle-clicking on the main canvas
will cancel editing the roomname.

Clicking on New, Load, or Save will cancel editing the roomname as well.

Clicking on the trinket or crewmate locator buttons will cancel editing
the roomname, too.

Clicking on "Show all" will cancel the roomname edit, and if you start
editing the roomname while the tiles picker is up, the tiles picker will
be closed.