Fix being unable to tab language dialog in level select
You could tab in the dialog if you did it in Ved options, but for some
reason you couldn't do it in level select. Well, that was because it was
getting caught on the fact that input was being taken. Dialogs use their
own separate system for taking input and don't use the 'takinginput'
global var anyway, so we can safely tell Ved to ignore 'takinginput' if
a dialog is open.