Holding right click on start point tool now only finds start point once
There's two ways to seek the start point, ctrl+P and right-clicking the
start point tool. Right-clicking the start point tool would previously
repeat the action while you're holding the right mouse button, which is
not noticeable but it's kinda useless and it does spam messages in the
console if you have it open.

[pre03]