Add trinkets/crewmate/start point locator on the map
As suggested by Dav in !29.

The trinket, crewmate, and start point tools are shown on the left-hand
side of the map screen now. You can click on them, and they will show
where each trinket/crewmate/start point is on the map, while also
darkening it beforehand for visual clarity.

The smaller screen is supported too - you just have to hold Ctrl if you
want to see the tools or click on them.

Of course, you can use the tool keybinds as well. I didn't draw the tool
keybinds because I don't think you need to be reminded what they are.
You can right-click on the start point icon to go to the start point
room, or press Ctrl+P. You can scroll through them using the scrollwheel
or < and > (you don't have to use Ctrl/Shift+Scrollwheel, and
Ctrl/Shift+</> doesn't affect the scrolling). Also, the scroll direction
will be reversed if you have s.flipsubtoolscroll on.

If there are three or less trinkets within a room, Ved will draw a
trinket for each trinket. Otherwise, it'll just draw the number of
trinkets in the room instead. Same deal if there are three or less
crewmates within a room, except Ved will also draw the color of each
crewmate. Ved will draw one player always if there is at least one start
point, no matter what, and whichever start point it picks, it will draw
it with its direction as well. (We assume that trinkets are 16x16 pixels
and crewmates and players 12x21 pixels, and also assume that no one's
crazy enough to make those sprites bigger than that.)

This also works with the new room shifting feature, which needs special
code to be programmed for, or otherwise the trinkets/crewmates would
just stay in place visually without getting shifted.