Add support for reloading language in-place
If something is already `ved_require()`d, we unload it by doing
something like `package.loaded.<module name> = false` and then
`ved_require()` it again.

The text saying you have to reload Ved for the language change to take
effect is no longer used, and the rest of the fields in the Language
dialog have been shifted up accordingly.