Add support for resetting corresponding map rooms
map_correspondreset() has been added, along with constants for
indicating which attribute of a map room is dirty.

If a room is row-dirty, map_correspondreset() will reset all rooms
referencing those rows. If a map is property-dirty,
map_correspondreset() will reset all rooms referencing those properties.
If a room is entity-dirty, it will simply reset it without resetting
other rooms. However, nothing in Ved currently marks rooms as
entity-dirty, because we don't draw entities on the map yet.

I've tried to limit the amount of room resets in a few different ways
(by making sure in map_correspondreset() we don't reset the same room
over and over again, or failing even that, by checking that we haven't
already reset the room in map_resetroom()), but it seems like Ved is
just reset-happy.

For example, in switchtileset() and switchtilecol(), Ved records that a
room's properties have changed into the undo/redo stack. This wouldn't
be a problem, but apparently it's also decided to keep a copy of the
current tiles of that room at that time - meaning if you undo a property
change, it will call roomdata_set() to change those tiles back, causing
unnecessary corresponding resets, and even moreso if the room in
question is one of the rooms who has a row in the y:1>=21 repeated rows
rooms!