Add autodetecting language based on OS language
Now there's a new autodetect option in the Language dropdown list. If
you select it, the language will be autodetected based on your operating
system's language every time Ved starts up, and every time you click OK
on the language dialog.

For macOS and Linux, we use the LANG environment variable. For Windows,
we use GetUserDefaultUILanguage().