Display translucent pixels like how VVVVVV does it
VVVVVV has a bit of buggy support regarding translucent pixels, that is,
pixels whose opacity is between 0% and 100% and not exactly 0% and not
exactly 100%. I mean, in the in-game editor, translucent pixels will
look fine - until you start playtesting, or open the level in the actual
game! Then again, this game seems to have problems with misleading
in-game editor visuals anyway (enemy color in Space Station tilecol -1,
the roomnames of (1,21) and (21,20) being the last editor message even
though they don't show up in-game, etc.).

While writing this code, I forgot to preserve alpha for pixels that were
completely transparent, and found some "hidden" non-black transparent
pixels in the mix - most notably, the white-outlined extra Lab tiles in
tiles2 were meant to have filling at one point or another! And tile 713
in tiles2 (the "background" of Lab and Warp Zone tilesets) is a
completely filled unused tile that just happened to be opacity 0 for
some reason, further adding to my conclusion that for some technical
reason they needed to put a tile for the background tool of those
tilesets, and just decided to make it this one tile. Makes me wonder
what kind of photo editor the devs used, or better yet, why no one has
discovered this until now.