Add a border to the main canvas
So if there is black on the sides, at least you can tell when the room
ends visually. Plus it's very aesthetically pleasing.

Since the interface is so tight, I've had to shrink the width of the
left panel's black background (both in normal window size and
s.smallerscreen size) by 1 pixel, move the right panel's black
background 1 pixel to the right, and move the "T:"/"C:"/"E:" text to the
right by 2 pixels. I don't think this causes any problems.