Fixed "plugin error" screen issues
The plugin error screen has been a bit neglected; due to an oversight
its background was black in LÖVE 0.10 and up (and even in 0.9, losing
focus would make the background black as well), its alignment was off
for some reason, and in LÖVE 11 and up it would misalign the 8px font
until a different font was displayed (usually pretty quickly when
switching to a different state/screen). This commit fixes all these
issues.

[pre03]