Stop double-counting commands in tab autocompletion
Previously, it was doublecounting commands like `delay`. The list of
known commands are split into `knowncommands` and
`knowninternalcommands`, which are simplified and internal respectively;
and when it tries to find a match it looks through both, starting with
`knowncommands` - but when it looks through `knowninternalcommands`
right after, it used to not check to make sure the command wasn't
already in the list of matches. So, in the example of `delay`, it would
add it once when it looked through `knowncommands`, and then add it
again when it looked through `knowninternalcommands` - then say "there
are two matches, this must be ambiguous" even though there is really
only one match.