Call temp_print_override() again when changing s.usefontpng
Otherwise, the print functions won't get properly hijacked in some
cases.

Of course, since we're calling `temp_print_override()` again when it was
originally meant to be used once during startup, we have to make sure
not to redefine `love.graphics.print11` and `love.graphics.printf11`
again, or else we'll end up crashing with a stack overflow.
Additionally, `font_skipnextreset` was removed since the thing it was
meant to fix was already properly fixed earlier in
8b13095119c97f4d1959e7bf7ebc678419d0b506, and it only causes issues now
when reloading or unloading font.png in-place.