DF.FILES: Make selection also accept name field
This means that you can click on a filename, and it'll keep being
highlighted once you move your mouse off of it. And you can type a
filename in manually to the "Name:" field and it'll get highlighted
correspondingly as well. And yes, I checked just now, and characters to
the left of the cursor are not ignored when you do this.