Show the Enter hotkey for Close, Save, Load, and Quit
This in order of their priority, following OK, with the ones to the left
having more priority than the ones to the right.

This does mean that Enter will be shown instead of S for Save - but I
feel like we shouldn't have multiple hotkeys for the same button anyway,
to minimize confusion even if we didn't have F9 hotkeys.

And also for some reason pressing S for Save only works if the dialog
has 0 fields at all (meaning it's absolutely blank apart from the
message), which is really weird. `git blame` traces this back to
511fc9388b03bc96be7ac0a295fedcdc82807205. I'm sure it was done for SOME
reason, but that commit doesn't explain what.

This means that S only works in the 'confirm quit with dirty level'
dialog but not on the map rendering dialog, but Enter works for both
regardless. So I'm glad this means Save will be showing Enter instead of
S, because Enter is clearly better.